Debugging
Debugging Print:
This commit is specifically to show these fun debugging skills via cout and the fun I have had trying to figure out why my algorithm was not working.
First I thought it might be the insert, but it seems to be the print function that has the issue as I have access to all of the pointers parents and values
going back up the line in correct order.