Amazing
When timing the print method and also viewing the total numbers sinse
randoms can appear more than once and the insert function ignores them
I was able determine a fairly steady amount of numbers that actually get inserted
into the binary tree. I ran a simple forloop with that number as its
conditional stop to see if it was faster and it is not faster. Amazing